Rename testing-unit-common module to testing-junit-common
