+++ testing-junit-common/src/main/java/com/bumble/appyx/testing/junit/common/util/DummyRoutingSource.kt
@@ -1,4 +1,4 @@
-package com.bumble.appyx.testing.unit.common.util
+package com.bumble.appyx.testing.junit.common.util
 
 import com.bumble.appyx.core.routing.BaseRoutingSource
 import com.bumble.appyx.core.routing.RoutingElements
