+++ testing-junit-common/src/main/java/com/bumble/appyx/testing/junit/common/util/TestIntegrationPoint.kt
@@ -1,4 +1,4 @@
-package com.bumble.appyx.testing.unit.common.util
+package com.bumble.appyx.testing.junit.common.util
 
 import com.bumble.appyx.core.integrationpoint.IntegrationPoint
 import com.bumble.appyx.core.integrationpoint.activitystarter.ActivityStarter
