+++ testing-junit5/build.gradle
@@ -37,8 +37,7 @@ android {
 }
 
 dependencies {
-    api project(':testing-unit-common')
-
+     api project(':testing-junit-common')
     implementation libs.junit.api
     implementation libs.androidx.arch.core.testing
     implementation libs.kotlin.coroutines.test.jvm
